testsuite/coverage.sh: Add "pragma: py3-only" to coverage exclude lines
As part of the transition to Python 3.x, we will start adding some code
which should only be executed when running the hooks using Python 3.x.
In the meantime, until the testsuite gets transitionned to Python 3.x,
we still want the coverage report for Python 2.x to remain clean, so
this commit adds a new exclude line to the default one: Any line with
"pragma: py3-only" should now be excluded from the coverage report.

As for the future, what we will do when the transition to Python 3.x
actually happens, is change the config again to exclude lines marked
"py2-only" (or remove the Python-2.x-only code directly, if that makes
better sense).

TN: U530-006
Change-Id: I67a1fe27dbfe614c1dba5176d78f1d8d1fe9cd35