git.diff_tree: Remove support for **kwargs
I noticed while working on the transition to Python 3.x that
this argument is not actually used, so this commit removes it.
It makes the API for this function simpler and also helps avoid
some pitfalls (as shown by one of the limitations we documented
that this commit removes).

TN: U530-006
Change-Id: I5ced7c6a979c5381d35498b06f1af8c569a3f9ed