delete_tag_not_allowed: Rename test_xxx method to match testing objective
I noticed while looking at this testcase that the name of the test_xxx
method used to perform the testing in this testcase did not match what
the testing was trying to do (likely because this testcase was created
by first copying an existing testcase and then modifying it). This commit
fixes the issue by renaming that method.

Found while working on: U530-006.

Change-Id: Icbddfd9bc2ea801df28b68219dcc09bfe5e9e62b