update_l_tag,update_tag: Remove code aimed at supporting old Git version
This commit removes a bit of code aimed at supporting the output
from an old version of Git, and does so by modifying the .out
attribute of the Run object we got from running Git. Unfortunately,
this won't work when we switch to Python 3.x and E3's Run class,
as the "out" attribute has become a property instead, so we cannot
modify its value. Since the version of Git we're trying to support
is very old at this point (latest release on 1.7 branch was Oct 2012),
we don't really need to keep that code anymore.

Found while working on U530-006.

Change-Id: I0ad83b00bf1d9e46ab67fd9bc50400f75308f32b