Remove most calls to unittest.TestCase.assertTrue in our testcases
TestCase.assertTrue does not bring anything compared to just using
the assert statement. So this commit uses the following perl command
to automatically rewrite all calls which fit in a single line
(which is most of our calls, but not all of them):

    $ cd testsuite/tests
    $ perl -pi -e 's/\S*\.assertTrue\((.*)\)\s*$/assert \1/' \
        **/run_test.py

This is preparation work for the transition of the testsuite out
of gnatpython (Python 2.x only), and relying on pytest instead,
which will allow us to run the testsuite using Python 3.x.

There are still a handful of calls to assertTrue due to them
being spread over more than one line. Those will be treated in
a separate commit to help with review, as those will be handled
manually.

TN: U530-006
Change-Id: I58daa2d6027e82edf5dc664159d08d9d9eaf2d7b