Remove all remaining calls to unittest.TestCase.assertTrue
This commit removes all the calls to assertTrue that were still
left. In this commit, all the changes were made by hand.

This is preparation work for the transition of the testsuite out
of gnatpython (Python 2.x only), and relying on pytest instead,
which will allow us to run the testsuite using Python 3.x.

TN: U530-006
Change-Id: I0f1925341954ff60a01149dd2075f87dfa3e573e