Rename "self" parameter to "testcase" in all our testcases
This is another preparation patch towards the transition of
the testsuite framework to pytest (as part of the transition
to Python 3.x), where we will use a fixture called "testcase"
to replace the TestCase object (self).

To facilitate the transition as well as the review of that (future)
transition, this commit renames the "self" parameter into "testcase".
Note that the change was performed entirely automatically thanks to
the following perl command:

    perl -pi -e 's/\bself\b/testcase/g' **/run_test.py

It's a departure from the convention that the object for which
a method is called be called "self", but it's only a temporary
situation; once we transition the testsuite to pytest, we will
no longer have classes -- just simple functions.

TN: U530-006
Change-Id: Id9cf32c8a8f1ed0c2fd4236f303c8fbd7b4894c1