Introduce a TestCase.run "method" (for transition to pytest)
This commit introduces a new run attribute in our testsuite's
TestCase class as a way to allow testcases to call this method
instead of instantiating the Run class. All testcases are then
modified to use this method, by running the following perl
command:

    $ cd testsuite/tests
    $ perl -pi -e 's/\bRun\(/testcase.run(/g' **/run_test.py

The testcases are then modified to remove the unused import using
the following perl command:

    $ perl -pi -e 's/\s+Run,//g' **/run_test.py

The goal is to help the transition to pytest, where the testcases
will be using the "run" method to call external commands such as
Git. The difference is that the pytest version of this method
will do more than being a simple wrapper (e.g. it will take care
of the default working directory).

TN: U530-006
Change-Id: I1823ef4331d8d16da6cf5f4a4efdbb667c68848d