TestCase: Add support for unit testing via script as separate process
This commit adds a new method to the TestCase class used by
the testsuite whose goal is to facilitate the calling of a script
which can then perform unit testing as a separate process from
the testcase's process.

This is preparation work for the transition out of gnatpython
(Python 2.x only) and towards pytest, where all the testing is done
within the same process. When that happens, changing globals such as
the environment should not be done, as the changes would cause
side-effects with other testcases that are run concurrently.

The "logger_error" is then re-implemented using this new method.
This allows us to greatly simplify this testcase as we don't have
to work so hard at creating a scenario resulting in a call to
the "syslog" function. As it happens, the scenario that we used
for this testcase involves some code which we want to get rid of
(now OBE), and the few other locations where "syslog" is called
do not lend themselves to traditional testing (those are current
covered via unit testing too).

TN: U627-004 (remove support for $HOME/.no_cvs_check)
TN: U627-007 (avoid changing testsuite environment in testcase)
TN: U530-006 (transition to Python 3.x)
Change-Id: I8f47dc4bc877387d52edddbcde10dd14a49b9b6b