Remove support for $HOME/.no_cvs_check file
Support for this specific file is purely AdaCore-specific and
historical. Since then, we have introduced (much!) better ways
to support users who want to suppress checks for a given commit.
We know this feature hasn't been used for many years, so it is
time to remove support for it. Note that it wasn't even documented.

TN: U627-004
Change-Id: Ie67532158abc1d302a3d98e57835f15dadfe0817