alternate_style_checker: Rewrite to avoid modifying os.environ
This is preparation for switching this testsuite to pytest (as
part of the transition to Python 3.x). With pytest, the testcases
are run within the same process, so changing any part of the global
environment such as os.environ is bound to affect other testcases.

This commit therefore enhances this testcase to avoid doing this.
Instead, it uses a copy of os.environ which it then modifies
and then passes it to the "git" command being called.

TN: U530-006
Change-Id: Id121c9ca266bfe9268a45ec5deead2e6a4cf3f74