Introduce new TestCase.bare_repo_dir property and use it in testcases
The goal of this commit is to factorize the code in a few of
the testcases that computes the path to the testcase's bare
repository. That way, the reference to TEST_DIR, a global variable
that we're planning on eliminating in favor of a TestCase attribute,
gets factorized in the TestCase class, outside of the testcases.

Done while working on the testsuite transition to pytest (itself also
part of the transition to Python 3.x.

TN: U530-006
Change-Id: Ib47b02a8f9eccc54ad13332786e560203fb7b80e