Remove all calls to "cd(testcase.bare_repo_dir)" in testsuite
This commit is preparation work towards the transition to pytest
(part of the transition to Python 3.x). It removes all calls to "cd"
which are changing the working directory to the testcase's bare repo
directory. Those are unnecessary if we use testcase.run's "cwd"
parameter instead.

This helps us reduce the number of instances where a testcase changes
its global environment, something we know will be an issue once
we transition the testsuite to pytest.

TN: U530-006
Change-Id: Ie5fa009439390cf2b8027ab7536788b673edea1d