Make TestCase.run a method rather than a (callable) attribute
The purpose of this change is to allow us to change the semantics
when calling testcase.run, by implementing these slight semantic
changes in that new method.

The ulterior goal of this change is to change the directory we run
commands from when the "cwd" parameter is None, changing it from
the current working directory, to the the directory where the test
repository is used. From there, we'll then be able to remove most
of the calls to "cd", something we have to do as part of the transition
of this testsuite to pytest.

TN: U530-006
Change-Id: I8ea69363aac28e1cbc4886dd387a191fd0f4bd03