Introduce new TestCase.repo_dir property
The purpose of this change is to help avoid the same code being
duplicated in the testcases themselves.

No testcase is changed to use this, yet, but some followup commits
will.

Found while working on the transition to pytest, where we'll need
to avoid changing the global environment, which means we'll need
to be more explicit about directories (avoid changing the working
directory, and/or avoid assuming what the current working directory
is).

TN: U530-006
Change-Id: Ieb90bd3c1ae09a8261ac88322c87a2ab6f8ec35f