Enhance some testcases to avoid assumption about current working directory
This commit enhances some testcase to avoid assuming that the current
working directory is the testcase's non-bare repository. The goal
is to eventually remove the call to "cd" which changes that current
working directory, something we know is going to cause problems
when we transition the testsuite to pytest.

TN: U530-006
Change-Id: I1d956c89a00143bcdda784ea9972686a567de68d