TestCase._run: run command in non-bare repo directory by default
This is another preparatory step towards being able to remove
the remaining calls to "cd" in our various testcases (something
that we cannot do once we transition to pytest, as it changes
the global environment).

Since the vast majority of commands being run by our testcases
are git commands to be run inside the non-bare repository,
this commit adjusts the TestCase._run method such that, if
a specific cwd is not specified, then run the command from
the directory of the non-bare repository.

TN: U530-006
Change-Id: I00d0685e02da24d83f652b035d37479eff310b1f