Remove all calls to "cd" in all testcases
All these calls are now superflous, so this commit removes them.

Note that this change was made entirely automatically, using
the following command:

    $ cd testsuite/tests
    $ perl -i -0pe 's/^\s*cd\([^\n]*\n+//gms' */run_test.py

Translation: Find all lines that start with spaces, followed by "cd(",
and then delete everything, including empty lines after it (if any).

The following command was then executed to automatically remove
all imports of "cd", which are now superfluous:

    $ cd testsuite/tests
    $ perl -i -e 's/ cd,//' */run_test.py

This helps our goal of avoiding changes to the global environment,
which is a requirement before we can transition the testsuite to
pytest.

TN: U530-006
Change-Id: Ia84162550648e622d8ab2d1b3210c8c4eeb0219b