testsuite/lib/support.py: Remove import of "cd"
This import is no longer used by any testcase, and can be removed.
This ensures we don't re-introduce calls to this function in the future.

This is part of the transition of the testsuite to pytest, where
testcase should not be changing the global environment (such as
the current working directory).

TN: U530-006
Change-Id: I13b0f4882295ad5980c92c0ee726d48614448284