testsuite: Replace the "TEST_DIR" global by a TestCase.work_dir property
This is another preparation patch towards the transition to pytest where
testsuite support code is accessed through fixtures, rather than imports.
Having the equivalent of this information available as a property rather
than a global means we have access to this information though the "testcase"
object, something that works both with the current testsuite framework
and with the future pytest-based testsuite framework.

Note that the changes in the testcases themselves have been semi-automated.
I started with...

    $ cd testsuite/tests
    $ perl -pi -e 's/TEST_DIR/testcase.work_dir/g' */run_test.py

And then fixed up by hand the 30 testcases or so which had failures
in them. For the most part, it was about removing the import of
TEST_CASE, which had been replaced by an import of testsuite.work_dir.

TN: U530-006
Change-Id: I154dd58832dc10909623b869c488aa39edaf4c2d