git_show_ref: Avoid changing dict size while iterating over it
This is another Python 3.x preparation patch, which fixes a situation
where we change a dictionary size while iterating over it, something
that became apparent while exercizing the hooks with Python 3.x.

This commit avoids this by rewriting this part of the function
in a way that aboves the dictionary iteration entirely.

TN: U530-006
Change-Id: Ia22b28919ac2dabc3e22ed1b3652e14d799e566b