cannot_find_style_checker/run_test.py: Use "format" method instead of "%"
This commit slightly rewrites the way we format the expected output
of our of our tests, by using the "format" method instead of using
the "%" operator. It doesn't make much of a difference, at the moment,
but this will become handy when we transition to Python 3.x, where
the "No such file or directory" error now includes the path of
the file/directory that's missing.

TN: U530-006
Change-Id: I748ca76d411cfdf79dd33402d1297e7372decc38