emails.py: Mark a couple of code blocks as being Python-2.x only
This commit adds a couple of "# pragma: py2-only" comments to
a couple of code blocks which are only expected to be run when
the hooks are tested with Python 2.x (these two code blocks are
conditioned on the version of Python being less than 3).

This will help us manage the transition to Python 3.x until we are
able to drop support for Python 2.x. That way, we can run coverage
analysis with both Python 2.x and Python 3.x, and get the Python 3.x
coverage analyzer to ignore those blocks we know we cannot cover
with Python 3.x.

Once the transition to Python 3.x is over, we will remove those code
blocks.

TN: U530-006
Change-Id: I44f1cd883c3fdf4e487e1e553158517e721416df