encode input and decode output when calling hooks.mailinglist script
This is another preparation patch for the transition to Python 3.x.

The script's input needs to be encoded when called, and its output
needs to be decoded into a string for us to process it.

For input encoding, the same approach as for decoding is taken:
In order to make progress towards Python 3.x support while at
the same time preserving support for Python 2.x, we introduce
a new function "encode_utf8" which only performs the encoding
on Python 3.x. With Python 2.x, the function just returns the string
unmodified.

Change-Id: Ieb47d32c756405cdd0d300254e8cd7c8c3db50b5
TN: U530-006