updates/emails.py: encode input and decode output when calling filer
This is another preparation patch for the transition to Python 3.x.
With Python 3.x, we need to make sure that the input used when
calling the filer cmd is converted to a byte string. We also
then need to make sure that the script's output is decoded into
a string.

Change-Id: I324410dd5c9b1e811252803b854d0f06ca65435d
TN: U530-006