updates/sendmail.py: encode input and decode output when calling sendmail
This is another preparation patch for the transition to Python 3.x.
With Python 3.x, we need to make sure that the input used when
calling sendmail is converted to a byte string. We also then need
to make sure that the script's output is decoded into a string
when printing it.

Change-Id: I1b792638fb77c8d1b4ee2197b29b63922e0fe211
TN: U530-006