testsuite: Update expected output to Git 2.32 (with 2.25 compatibility)
This commit upgrades the testsuite's expected output to the output
produced by Git version 2.32.0. However, in an effort to make it easier
for users to run the testsuite in other environments, a bit of testsuite
infrastructure is also introduced whose role is to adapt the expected
output to the version actually available during testing. The initial
version of this adaptor allows us to maintain testsuite compatibility
with Git version 2.25.x.

Change-Id: I6fc353c92c6e680c1a6155d4f961c0b1243e4133