update/post_receive: Do nothing if reference in hooks.ingore-refs
We noticed that Gerrit's ref-updated hook was reporting some
errors when Gerrit makes updates to some of its internal references.
For instance:

    | Traceback (most recent call last):
    |   File "./hooks/post_receive.py", line 126, in <module>
    [...]
    | errors.InvalidUpdate: ('Unable to determine the type of reference
    |    for: refs/changes/01/104201/meta', '',
    [...]

This commit avoids this error by making sure that we properly
ignore all references that match the hooks.ignore-refs config
(note that, by default, hooks.ignore-refs includes matches for
those special Gerrit references).

A new testcase is added to verify that this is the case.

Note that the minor reformatting in maybe_update_hook comes from
black, which adds the extra coma only when the changes made in
this commit are made as well. This is why this reformatting change
is bundled with this commit.

Change-Id: Id94129ad1a9d84ccba6acafef3d8e98f62ec0239
TN: UA21-052