fast_forward.py: Fix arguments in call to warn from InvalidUpdate handler
This is a change that comes from the transition to Python 3.x.
The only way it was detected by the current testsuite is that
the line of code located just after the call to warn appeared
as not being covered, which was odd, since the function "warn"
is supposed to always return. To investigate this further,
I simply added an invalid line of code right after the call
to "warn"...

    except InvalidUpdate as E:
        warn(*E)
        invalid()

... and re-ran the testsuite: clean results (and invalid is
not covered).

Looking at the code more closely allowed me to realize that,
with Python 3.x, we need to pass *E.args, not *E. As a result,
we get an exception when calling warn as follow:

    TypeError: __main__.warn() argument after * must be an iterable,
    not InvalidUpdate

The reason we're not seeing this error in the testsuite is
because the issue is in a part of the file which only gets
executed when that file is used as a Python script, something
that the git-hooks don't do (the git-hooks imports the module
and calls the check_fast_forward function). Understanding this
then allowed us to track the testcase that tests the script,
and see that it was checking the script's status code, but
not its output. This commit enhances the testcase accordingly.

Change-Id: Icb92d82be6e108c6f05d84f0df5b787b97783082
TN: U530-006