conftest.py: Simplify Run.cmd_out (remove support for old Git versions)
This commit simplifies the implementation of the cmd_out property
in the testsuite's Run class by removing some output substitution
code designed to run the testsuite against versions of Git which
are now very old (the last 1.6 and 1.7 releases are from 2010 and
2012). We no longer want to support such old versions, so this code
is no longer necessary.

Note that one testcase had to be updated as a result of this change.
This is expected, as the old code was masking (stripping) some output
printed by current versions of Git.

Change-Id: I9b6c294478e4ef65ab346b7e9a48d0b333c3c6ba