Add support for adding a warning banner when sending emails
This warning banner is triggered by the presence of an environment
variable to be set by the user prior to manually calling the
post_receive hook. This banner is then inserted at the beginning
of the email body in order to warn all readers of that email
that the email was not automatically sent at the time of push,
but rather at a later time.

TN: UC02-038
Change-Id: I42dece88dd0619df33adcc44d7adfc3ccd63a162