Fix "to allow us to allow us" duplication in script documentation
Pierre noticed during review of an earlier patch that
a script used by the associated testcase had some documentation
saying "to allow us" twice. I fixed the issue for the new testcase
introduced in the previous commit, but I knew that this script
was a copy of a script which is quite standard across the testsuite.
This commit fixes the issue for all the other instances of that
same script (all 111 of them ;-)).

For the record, the change was performed entirely manually
using the following command:

    $ perl -pi -e 's/to allow us to allow us/to allow us/g' \
        `git grep 'to allow us to allow us' | cut -d: -f1 | sort -u`

Change-Id: Ia532067b1bcdac22f72b874d67f0f210fdf83862
TN: V103-002