exclude ignored refs from git-notes commit emails information
A recent change enhanced the emails being sent for git notes
to also provide the list of branches (technically, the references)
which contain the annotated commit (S731-057). For instance:

    | Subject: [notes][repo/branch] Annotated commit subject
                       ^^^^^^^^^^^
And also this section in the "Diff:"

    | For the record, the references containing the annotated
    | commit above are:
    |
    |     refs/heads/branch

However, there is a small hole in the implementation. I forgot
to take into account the "hooks.ignore-refs" config. As a result,
for repositories hosted on Gerrit, the emails mention references
which are internal to Gerrit.

    | Subject: [notes][repo/master,(refs/changes/67/108467/1)] subject
                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^

... and ... in the email:

    | For the record, the references containing the annotated commit
    | above are:
    |
    |     refs/changes/67/108467/1
    |     refs/heads/master

This commit fixes this oversight.

Change-Id: I4e21d5c906e94c01b650615282258cc7b4fb81d9
TN: S731-057 (ticket introducing this feature)
TN: V105-012 (ticket opened to fix the oversight)