testsuite/README.md: Small formatting adjustments
This commit adjusts the formatting used in the blocks providing
the command to run the testsuite, by using the "console" identifier,
as is already done in other parts of the git-hooks READMEs.
In addition, it removes the identation that was inserted at
the start of each command, as this indentation isn't consistently
followed and doesn't really serve any real purpose.

Change-Id: I3c2a0b84ff3b1c0d0a30d0ec5371681cb64d3d17
TN: V104-006