replace distutils.version by packaging.version
distutils.version is now deprecated, so transition to packaging.version
instead. Unfortunately, the Version object from packaging.version
is not as convenient to use as distutils.version, as the comparison
operators do not support the use of strings as the righ-hand-side.
So users of those objects are updated accordingly.

Change-Id: I72aeb40050e5da25ead616b3d9475218ba88e403
TN: V304-012