treat all refs/users/.* references as internal and to be ignored
We noticed that Gerrit can create some references using the following
naming scheme:

    refs/users/<last-2-digits-of-userid>/<userid>

When it does, this causes the hooks to fail with a message indicating
that it is unable to determine the type of reference it is, because
it doesn't recognize it.

This commit changes the hooks.ignore-refs default configuration
to ignore all references whose name start with "refs/users/".
We could be stricter in our matching, but I think this would be
slightly overkill. And meanwhile, we can see that this is not
the first time we notice that references in the "refs/users/"
namespace where created by Gerrit (e.g. we were already ignoring
references matching "refs/users/.*/edit-.*" before). So, being
being more relaxed here helps us avoid having to change the config
again in the future should Gerrit decide to expand on its use of
this namespace.

Change-Id: I3196cfc833ac94473d4b5fcdca773ded4ced24cf
TN: V218-013